requirements.txt
openai
python-git
streamlit
tiktoken
python-dotenv
langgraph
loguru
fastapi
langchain_openai
uvicorn